The Adaptation slider is properly reset when unchecking the Preconditioning profile checkbox

+++ DisplayCAL/wxTestchartEditor.py
@@ -2434,7 +2434,7 @@ END_DATA"""
     def tc_precond_handler(self, event=None):
         setcfg("tc_precond", int(self.tc_precond.GetValue()))
         self.tc_adaption_slider.SetValue(
-            (1 if getcfg("tc_precond") else defaults["tc_adaption"]) * 100
+            int((1 if getcfg("tc_precond") else defaults["tc_adaption"]) * 100)
         )
         self.tc_adaption_handler(self.tc_adaption_slider)
         self.tc_algo_handler()
